official/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
boost-cpp=1.74.0=h9f4b32c_8
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.6.15=h5b45459_0
certifi=2022.6.15=py39hcbf5309_0
cgal-cpp=5.5=h3c2f9b9_0
colorama=0.4.5=pyhd8ed1ab_0
curl=7.83.1=h789b8ee_0
double-conversion=3.2.0=h0e60522_0
eigen=3.4.0=h2d74725_0
expat=2.4.8=h39d44d4_0
ffmpeg=4.3.1=ha925a31_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=hce3cb01_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freeimage=3.18.0=h064cf08_10
freetype=2.12.1=h546665d_0
gl2ps=1.4.2=h0597ee9_0
glew=2.1.0=h39d44d4_2
hdf4=4.2.15=h0e5069d_4
hdf5=1.12.1=nompi_h2a0e4a3_104
icu=68.2=h0e60522_0
ifcopenshell=v0.7.0a4=py39hf351b84_0
imath=3.1.5=h12d4b20_0
jpeg=9e=h8ffe710_2
jsoncpp=1.9.5=h2d74725_1
jxrlib=1.1=h8ffe710_2
krb5=1.19.3=h1176d77_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libclang=11.1.0=default_h5c34c98_1
libcurl=7.83.1=h789b8ee_0
libdeflate=1.12=h8ffe710_0
libiconv=1.16=he774522_0
libnetcdf=4.8.1=nompi_h1cc8e9d_102
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h1d00b33_4
libraw=0.20.2=hee1bdec_1
libssh2=1.10.0=h680486a_3
libtheora=1.1.1=h8d14728_1005
libtiff=4.4.0=h2ed3b44_1
libwebp-base=1.2.4=h8ffe710_0
libxml2=2.9.14=hf5bbc77_4
libzip=1.9.2=hfed4ece_1
libzlib=1.2.12=h8ffe710_2
loguru=0.6.0=py39hcbf5309_1
lz4-c=1.9.3=h8ffe710_1
mpfr=4.1.0=h8d14728_1
mpir=3.0.0=he025d50_1002
occt=7.6.2=h8294837_3
openexr=3.1.5=hab3b255_0
openjpeg=2.5.0=hc9384bd_1
openssl=1.1.1q=h8ffe710_0
pip=22.1.2=py39haa95532_0
proj=9.0.0=h1cfcee9_1
pugixml=1.11.4=h0e60522_0
python=3.9.12=h6244533_0
python_abi=3.9=2_cp39
qt=5.12.9=h5909a2a_4
rapidjson=1.1.0=ha925a31_1002
setuptools=61.2.0=py39haa95532_0
sqlite=3.38.5=h2bbff1b_0
tbb=2021.5.0=h2d74725_1
tbb-devel=2021.5.0=h2d74725_1
tzdata=2022a=hda174b7_0
utfcpp=3.2.1=h57928b3_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
vtk=9.1.0=qt_py39h1ab545e_208
wheel=0.37.1=pyhd3eb1b0_0
win32_setctime=1.1.0=pyhd8ed1ab_0
wincertstore=0.2=py39haa95532_2
xz=5.2.6=h8d14728_0
zlib=1.2.12=h8ffe710_2
zstd=1.5.2=h6255e5f_4
